[#593] Clean up implementation
