+++ Cargo.toml
@@ -27,7 +27,6 @@ members = [
 
     "iceoryx2-services/discovery",
 
-    "iceoryx2-tunnels/common",
     "iceoryx2-tunnels/end-to-end-testing",
     "iceoryx2-tunnels/zenoh",
 
@@ -83,7 +82,6 @@ iceoryx2-ffi = { version = "0.6.1", path = "iceoryx2-ffi/ffi" }
 iceoryx2-ffi-python = { version = "0.6.1", path = "iceoryx2-ffi/python" }
 iceoryx2-ffi-macros = { version = "0.6.1", path = "iceoryx2-ffi/ffi-macros" }
 iceoryx2-services-discovery = { version = "0.6.1", path = "iceoryx2-services/discovery"}
-iceoryx2-tunnels-common = { version = "0.6.1", path = "iceoryx2-tunnels/common" }
 iceoryx2-tunnels-end-to-end-testing = { version = "0.6.1", path = "iceoryx2-tunnels/end-to-end-testing" }
 iceoryx2-tunnels-zenoh = { version = "0.6.1", path = "iceoryx2-tunnels/zenoh" }
