--- iceoryx2-tunnels/common/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "iceoryx2-tunnels-common"
-description = "iceoryx2: common components of tunnels"
-categories = { workspace = true }
-edition = { workspace = true }
-homepage = { workspace = true }
-keywords = { workspace = true }
-license = { workspace = true }
-readme = "../README.md"
-repository = { workspace = true }
-rust-version = { workspace = true }
-version = { workspace = true }
-
-[lib]
-name = "iceoryx2_tunnels_common"
-path = "src/lib.rs"
-
-[dependencies]
-iceoryx2 = { workspace = true }
-
-[dev-dependencies]
-
