+++ iceoryx2/src/service/port_factory/notifier.rs
@@ -46,7 +46,7 @@ pub struct PortFactoryNotifier<'factory, Service: service::Service> {
     default_event_id: EventId,
 }
 
-unsafe impl<'factory, Service: service::Service> Send for PortFactoryNotifier<'factory, Service> {}
+unsafe impl<Service: service::Service> Send for PortFactoryNotifier<'_, Service> {}
 
 impl<'factory, Service: service::Service> PortFactoryNotifier<'factory, Service> {
     pub(crate) fn new(factory: &'factory PortFactory<Service>) -> Self {
